Move text interface to the anntation flder
