+++ frontend/src/components/annotation/text_annotation/render_elements/fast_label_menu.vue
