+++ frontend/src/components/annotation/text_annotation/render_elements/selection.vue
