+++ frontend/src/components/annotation/text_annotation/render_elements/text_context_menu.vue
