+++ frontend/src/components/annotation/text_annotation/text_hotkeys.vue
