+++ frontend/src/components/annotation/text_annotation/text_sidebar.vue
@@ -131,11 +131,11 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import instance_detail_list_view from "../annotation/image_annotation/instance_detail_list_view.vue";
-import button_with_menu from '../regular/button_with_menu.vue';
-import label_select_only from '../label/label_select_only.vue'
-import attribute_group_list from '../attribute/attribute_group_list.vue'
-import global_attributes_list from '../attribute/global_attributes_list.vue'
+import instance_detail_list_view from "../image_annotation/instance_detail_list_view.vue";
+import button_with_menu from '../../regular/button_with_menu.vue';
+import label_select_only from '../../label/label_select_only.vue'
+import attribute_group_list from '../../attribute/attribute_group_list.vue'
+import global_attributes_list from '../../attribute/global_attributes_list.vue'
 
 
 export default Vue.extend({
