+++ frontend/src/components/annotation/text_annotation/text_utils/closest_token.ts
