+++ frontend/src/components/annotation/text_annotation/text_utils/draw_rects.ts
